Fix stop plugins on windows (#11)
Trying to stop a plugin on Windows lead to a dangling node process which prevented stopping the run completely. We now call the plugin-scripts directly to avoid such problems with yarn on Windows. If the watch script in package.json does not use plugin-scripts we start it with yarn as before as a fallback, knowing that stopping the plugin might not work on Windows.

Co-authored-by: Sebastian Sdorra <sebastian.sdorra@cloudogu.com>